backend/requirements.txt
attrs==22.2.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pytest==7.2.2
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
et-xmlfile==1.1.0
filetype==1.2.0
Flask==2.2.3
flask-sieve==2.0.1
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
openpyxl==3.1.2
Pillow==9.4.0
pydantic==1.10.7
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
urllib3==1.26.15
watchdog==3.0.0
Werkzeug==2.2.3